Clarify behaviour of 32-bit CSR writes to 64-bit registers
Counters like instret are defined to be 64-bit registers, even on RV32.
The terms "register" and "CSR" can ambigously refer to either the full
64-bit counter, or to the low or high 32 bits (such as instreth).

To reduce confusion, this patch adds clarifying text to describe the
expected behaviour of 32-bit CSR writes to 64-bit registers when these
registers can be updated as a side effect of instruction exectuion.